Set tlo-cpp options as normal variables
tlo-cpp can now handle this properly because the string_option() macro
won't create a cache entry if a normal variable with the same name is
already set and no cache entry with the name has been defined with -D.
Because a cache entry is not created, the normal variable binding won't
be removed. See CMake documentation on "set" command.