Add option to let program record sources of hashes
If the option is enabled and two different input text files contain the
same fuzzy hash, these hashes will not be considered duplicates and they
both will inserted into the HashComparisonMap for comparison. Also,
whenever a pair of similar hashes is found, the program will output
which input text file each of the hashes came from.